% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_melt_curve_functions.R
\name{calculate_drdt_plate}
\alias{calculate_drdt_plate}
\alias{getdRdTall}
\title{Calculate dR/dT of melt curves for of every well in a plate.}
\usage{
calculate_drdt_plate(platemelt, method = "spline", ...)

getdRdTall(platemelt, method = "spline")
}
\arguments{
\item{platemelt}{data frame describing melt curves, including variables
well, Temperature, Fluor (fluorescence value).}

\item{method}{to use for smoothing:

  "spline" default, uses smoothing spline stats::smooth.spline.

  "diff" base::diff for lagged difference}

\item{...}{other arguments to pass to smoothing method.}
}
\value{
platemelt with additional column dRdT.
}
\description{
dR/dT, the derivative of the melt curve (of fluorescence signal R vs
temperature T), has a maximum at the melting temperature Tm. A single peak in
this suggests a single-liength PCR product is present in the well.
}
\section{Functions}{
\itemize{
\item \code{getdRdTall}: 
}}

\seealso{
Other melt_curve_functions: 
\code{\link{calculate_dydx_1}()}
}
\concept{melt_curve_functions}
