% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_melt_curve_functions.R
\name{calculate_dydx_1}
\alias{calculate_dydx_1}
\title{Calculate dy/dx vector from vectors y and x}
\usage{
calculate_dydx_1(x, y, method = "spline", ...)
}
\arguments{
\item{x}{input variable, numeric vector, assumed to be temperature}

\item{y}{output variable, numeric vector of same length as x, assumed
to be fluorescence signal.}

\item{method}{to use for smoothing:

  "spline" default, uses smoothing spline stats::smooth.spline.

  "diff" base::diff for lagged difference}

\item{...}{other arguments to pass to smoothing method.}
}
\value{
estimated first derivative of y with respect to x, numeric vector of
  same length as y.
}
\description{
Used in tidyqpcr to calculate dR/dT for a melt curve of fluorescence signal R
vs temperature T.
}
\seealso{
Other melt_curve_functions: 
\code{\link{calculate_drdt_plate}()}
}
\concept{melt_curve_functions}
