% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deltacq.R
\name{calculate_normvalue}
\alias{calculate_normvalue}
\title{Calculate a normalized value for a subset of reference ids}
\usage{
calculate_normvalue(
  value_df,
  ref_ids,
  value_name = "value",
  id_name = "id",
  norm_function = median
)
}
\arguments{
\item{value_df}{data frame containing relevant columns.}

\item{ref_ids}{values of reference ids, that are used to calculated
normalized reference value.}

\item{value_name}{name of column containing values.
This column should be numeric.}

\item{id_name}{name of column containing ids.}

\item{norm_function}{Function to use to calculate the value to normalize by.
Default function is median, alternatively could use mean, geometric mean,
etc.}
}
\description{
This is used to calculate the normalized `cq` values for reference
`target_ids` (genes), to use in `delta_cq` calculation for each `sample_id`.
}
\details{
Also used to calculate the normalized `delta_cq` values for reference
`sample_ids`, to use in `deltadelta_cq` calculation for each `target_id`.
}
