% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_blank_plate}
\alias{create_blank_plate}
\title{Create a blank plate template as a tibble}
\usage{
create_blank_plate(WellR = LETTERS[1:16], WellC = 1:24)
}
\arguments{
\item{WellR}{Vector of Row descriptors, usually LETTERS}

\item{WellC}{Vector of Column descriptors, usually numbers}
}
\value{
tibble (data frame) containing all pairwise combinations of WellR and WellC, as well as individual Well names in its own columns. Both WellR and WellC are coerced to character vectors, even if WellC is supplied as numbers. Default value is a 384-well plate.
}
\description{
Create a blank plate template as a tibble
}
\examples{
create_blank_plate(WellR=LETTERS[1:2],WellC=1:3)
create_blank_plate(WellR=LETTERS[1:8],WellC=1:12)
}
