% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_colkey_4dilutions_mRTNT_in24}
\alias{create_colkey_4dilutions_mRTNT_in24}
\title{Create a 4-dilution column key for primer calibration}
\usage{
create_colkey_4dilutions_mRTNT_in24(Dilution = c(1, 1/5, 1/25, 1/125, 1,
  1), DilutionNice = c("1x", "5x", "25x", "125x", "-RT", "NT"),
  Type = c(rep("+RT", 4), "-RT", "NT"), BioRep = rep(c("A", "B"), each
  = 12, length.out = 24), TechRep = rep(1:2, each = 6, length.out = 24))
}
\arguments{
\item{Dilution}{Numeric vector of length 6 describing sample dilutions}

\item{DilutionNice}{Character vector of length 6 with nice labels for sample
dilutions}

\item{Type}{Character vector of length 6 describing type of sample (+RT, -RT,
NT)}

\item{BioRep}{Character vector of length 6 describing biological replicates}

\item{TechRep}{Character vector of length 6 describing technical replicates}
}
\value{
tibble (data frame) with 24 rows, and columns WellC, Dilution,
  DilutionNice, Type, BioRep, TechRep.
}
\description{
Creates a 24-column key for primer calibration, with 2x BioReps and 2x
TechReps, and 5-fold dilution until 5^4 of +RT; then -RT, NT controls. That
is a total of 6 versions of each sample replicate.
}
\examples{
create_colkey_4dilutions_mRTNT_in24()
}
\seealso{
Other plate creation functions: \code{\link{create_blank_plate}},
  \code{\link{create_colkey_6dilutions_mRTNT_in24}},
  \code{\link{create_colkey_6in24}},
  \code{\link{create_rowkey_4in16}},
  \code{\link{create_rowkey_8in16_plain}},
  \code{\link{display_plate}},
  \code{\link{label_plate_rowcol}}
}
\concept{plate creation functions}
