% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_colkey_6_in_24}
\alias{create_colkey_6_in_24}
\title{Create a 6-value, 24-column key for plates}
\usage{
create_colkey_6_in_24(...)
}
\arguments{
\item{...}{Vectors of length 6 describing well contents,
e.g. sample_id or target_id}
}
\value{
tibble (data frame) with 24 rows, and columns
well_col, prep_type, tech_rep, and supplied values.
}
\description{
Create a 24-column key with 6 values repeated over 24 plate columns.
Each of the 6 values is repeated over 3x +RT technical replicates and 1x -RT.
}
\examples{
create_colkey_6_in_24(sample_id=LETTERS[1:6])
}
\seealso{
Other plate creation functions: 
\code{\link{create_blank_plate}()},
\code{\link{create_colkey_4diln_2ctrl_in_24}()},
\code{\link{create_colkey_6diln_2ctrl_in_24}()},
\code{\link{create_rowkey_4_in_16}()},
\code{\link{create_rowkey_8_in_16_plain}()},
\code{\link{display_plate}()},
\code{\link{label_plate_rowcol}()}
}
\concept{plate creation functions}
