% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_colkey_6diln_2ctrl_in24}
\alias{create_colkey_6diln_2ctrl_in24}
\title{Create a 6-dilution column key for primer calibration}
\usage{
create_colkey_6diln_2ctrl_in24(
  Dilution = c(5^(0:-5), 1, 1),
  DilutionNice = c("1x", "5x", "25x", "125x", "625x", "3125x", "-RT", "NT"),
  Type = c(rep("+RT", 6), "-RT", "NT"),
  TechRep = rep(1:3, each = 8, length.out = 24)
)
}
\arguments{
\item{Dilution}{Numeric vector of length 8 describing sample dilutions}

\item{DilutionNice}{Character vector of length 8 with nice labels for sample
dilutions}

\item{Type}{Character vector of length 8 describing type of sample (+RT, -RT,
NT)}

\item{TechRep}{Character vector of length 8 describing technical replicates}
}
\value{
tibble (data frame) with 24 rows, and variables WellC, Dilution,
  DilutionNice, Type, BioRep, TechRep.
}
\description{
Creates a 24-column key for primer calibration, with 1x BioReps and 3x
TechReps, and 5-fold dilution until 5^6 of +RT; then -RT (no reverse
transcriptase), NT (no template) negative controls. That is a total of 8
versions of each replicate.
}
\examples{
create_colkey_6diln_2ctrl_in24()
}
\seealso{
Other plate creation functions: 
\code{\link{create_blank_plate}()},
\code{\link{create_colkey_4diln_2ctrl_in24}()},
\code{\link{create_colkey_6in24}()},
\code{\link{create_rowkey_4in16}()},
\code{\link{create_rowkey_8in16_plain}()},
\code{\link{display_plate}()},
\code{\link{label_plate_rowcol}()}
}
\concept{plate creation functions}
