% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_colkey_6diln_2ctrl_in_24}
\alias{create_colkey_6diln_2ctrl_in_24}
\title{Create a 6-dilution column key for primer calibration}
\usage{
create_colkey_6diln_2ctrl_in_24(
  dilution = c(5^(0:-5), 1, 1),
  dilution_nice = c("1x", "5x", "25x", "125x", "625x", "3125x", "-RT", "NT"),
  prep_type = c(rep("+RT", 6), "-RT", "NT"),
  tech_rep = rep(1:3, each = 8, length.out = 24)
)
}
\arguments{
\item{dilution}{Numeric vector of length 8 describing sample dilutions}

\item{dilution_nice}{Character vector of length 8 with nice labels for sample
dilutions}

\item{prep_type}{Character vector of length 8 describing type of sample (+RT,
-RT, NT)}

\item{tech_rep}{Character vector of length 8 describing technical replicates}
}
\value{
tibble (data frame) with 24 rows, and variables well_col, dilution,
  dilution_nice, prep_type, biol_rep, tech_rep.
}
\description{
Creates a 24-column key for primer calibration, with 1x biol_reps and 3x
tech_reps, and 5-fold dilution until 5^6 of +RT; then -RT (no reverse
transcriptase), NT (no template) negative controls. That is a total of 8
versions of each replicate.
}
\examples{
create_colkey_6diln_2ctrl_in_24()
}
\seealso{
Other plate creation functions: 
\code{\link{create_blank_plate}()},
\code{\link{create_colkey_4diln_2ctrl_in_24}()},
\code{\link{create_colkey_6_in_24}()},
\code{\link{create_rowkey_4_in_16}()},
\code{\link{create_rowkey_8_in_16_plain}()},
\code{\link{display_plate}()},
\code{\link{display_well_value}()},
\code{\link{label_plate_rowcol}()}
}
\concept{plate creation functions}
