% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_rowkey_4_in_16}
\alias{create_rowkey_4_in_16}
\title{Create a 4-value, 16-row key for plates}
\usage{
create_rowkey_4_in_16(...)
}
\arguments{
\item{...}{Vectors of length 4 describing well contents, e.g. sample_id or
target_id}
}
\value{
tibble (data frame) with 16 rows, and variables well_row, prep_type,
  tech_rep, and supplied values.
}
\description{
Create a 16-row key with 4 values repeated over 16 plate rows. Each of the 4
values is repeated over 3x +RT Techreps and 1x -RT.
}
\details{
This helps to create plate layouts with standard designs.
}
\examples{
create_rowkey_4_in_16(sample_id=c("sheep","goat","cow","chicken"))
}
\seealso{
Other plate creation functions: 
\code{\link{create_blank_plate}()},
\code{\link{create_colkey_4diln_2ctrl_in_24}()},
\code{\link{create_colkey_6_in_24}()},
\code{\link{create_colkey_6diln_2ctrl_in_24}()},
\code{\link{create_rowkey_8_in_16_plain}()},
\code{\link{display_plate_qpcr}()},
\code{\link{display_plate_value}()},
\code{\link{display_plate}()},
\code{\link{label_plate_rowcol}()},
\code{\link{make_row_names_echo1536}()},
\code{\link{make_row_names_lc1536}()}
}
\concept{plate creation functions}
