% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_rowkey_4in16}
\alias{create_rowkey_4in16}
\title{Create a 4-value, 16-row key for plates}
\usage{
create_rowkey_4in16(...)
}
\arguments{
\item{...}{Vectors of length 4 describing well contents, e.g. sample or
probe.}
}
\value{
tibble (data frame) with 16 rows, and variables WellR, Type, TechRep,
  and supplied values.
}
\description{
Create a 16-row key with 4 values repeated over 16 plate rows. Each of the 4
values is repeated over 3x +RT Techreps and 1x -RT.
}
\examples{
create_rowkey_4in16(SampleID=c("sheep","goat","cow","chicken"))
}
\seealso{
Other plate creation functions: 
\code{\link{create_blank_plate}()},
\code{\link{create_colkey_4dilutions_mRTNT_in24}()},
\code{\link{create_colkey_6dilutions_mRTNT_in24}()},
\code{\link{create_colkey_6in24}()},
\code{\link{create_rowkey_8in16_plain}()},
\code{\link{display_plate}()},
\code{\link{label_plate_rowcol}()}
}
\concept{plate creation functions}
