% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_rowkey_8in16_plain}
\alias{create_rowkey_8in16_plain}
\title{Create a plain 8-value, 16-row key for plates}
\usage{
create_rowkey_8in16_plain(...)
}
\arguments{
\item{...}{Vectors of length 8 describing well contents, e.g. sample or
probe.}
}
\value{
tibble (data frame) with 16 rows, and columns WellC, and supplied
  values.
}
\description{
Create a 16-row key with 8 values repeated over 16 plate rows. No other
information is included by default, hence "plain".
}
\examples{
create_rowkey_8in16_plain(Sample=c("me","you","them","him",
                                   "her","dog","cat","monkey"))
}
\seealso{
Other plate creation functions: \code{\link{create_blank_plate}},
  \code{\link{create_colkey_4dilutions_mRTNT_in24}},
  \code{\link{create_colkey_6dilutions_mRTNT_in24}},
  \code{\link{create_colkey_6in24}},
  \code{\link{create_rowkey_4in16}}
}
\concept{plate creation functions}
