% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_melt_curve_functions.R
\name{debaseline}
\alias{debaseline}
\title{Remove baseline from amplification curves (BETA)}
\usage{
debaseline(plateamp, maxcycle = 10)
}
\arguments{
\item{plateamp}{data frame with plate amplification data, including variables
well, Cycle, Fluor (fluorescence value), and Program. Assume program 2 for
amplification curves from Roche Lightcycler format data.}

\item{maxcycle}{maximum cycle value to use for baseline, before
amplification.}
}
\value{
platemap with additional columns per well:
\tabular{ll}{
  Base      \tab baseline /background value  \cr
  Signal    \tab normalized fluorescence signal, i.e. Fluor - Base
  }
}
\description{
Remove baseline from qPCR amplification curves
by subtracting median of initial cycles.
}
\details{
BETA function version because:

- assumes Roche Lightcycler format,
we should ideally replace "program==2" by something more generic?

- the rule-of thumb "baseline is median of initial 10 cycles"
has not been tested robustly
}
