% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{display_plate}
\alias{display_plate}
\title{Display an empty plate plan which can be populated with 
ggplot2 geom elements.}
\usage{
display_plate(plate)
}
\arguments{
\item{plate}{tibble with variables well_col, well_row.}
}
\value{
ggplot object; major output is to plot it
}
\description{
Display an empty plate plan which can be populated with 
ggplot2 geom elements.
}
\examples{
library(ggplot2)

# display empty plot of empty plate
display_plate(create_blank_plate_96well())

# display wells of empty plate filled by column
display_plate(create_blank_plate_96well()) + 
  geom_tile(aes(fill = well_col), colour = "black")

# display wells of empty 1536-well plate filled by row
display_plate(create_blank_plate_1536well()) + 
  geom_tile(aes(fill = well_row), colour = "black")

}
\seealso{
Other plate creation functions: 
\code{\link{create_blank_plate}()},
\code{\link{create_colkey_4diln_2ctrl_in_24}()},
\code{\link{create_colkey_6_in_24}()},
\code{\link{create_colkey_6diln_2ctrl_in_24}()},
\code{\link{create_rowkey_4_in_16}()},
\code{\link{create_rowkey_8_in_16_plain}()},
\code{\link{display_plate_qpcr}()},
\code{\link{display_plate_value}()},
\code{\link{label_plate_rowcol}()},
\code{\link{make_row_names_echo1536}()},
\code{\link{make_row_names_lc1536}()}
}
\concept{plate creation functions}
