% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_efficiency.R
\name{est_efficiency}
\alias{est_efficiency}
\title{Calibrate multiple probes by calculating detection efficiency and R squared}
\usage{
est_efficiency(ct_df, formula = Ct ~ log2(Dilution) + BioRep)
}
\arguments{
\item{ct_df}{a data frame with Ct data, 1 row per well

Must have columns Type, TargetID, Ct, Dilution. Only Type=="+RT" columns are used.}

\item{formula}{formula to use for log-log regression fit. 

Default value assumes multiple biological replicates Ct ~ log2(Dilution) + BioRep.
If only a single Biological Replicate, change to Ct ~ log2(Dilution).
If multiple SampleIDs, change to Ct ~ log2(Dilution) + SampleID. See ?formula for background}
}
\value{
data frame with columns: TargetID, efficiency, efficiency.sd, r.squared
}
\description{
See calibration vignette for example of usage.
}
\details{
Note efficiency is given in ratio, not per cent; multiply by 100 for that.
}
\seealso{
est_efficiency_1
}
