% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_efficiency.R
\name{est_efficiency_1}
\alias{est_efficiency_1}
\title{Calibrate probes by calculating detection efficiency and R squared}
\usage{
est_efficiency_1(ct_df_1, formula = Ct ~ log2(Dilution) + BioRep)
}
\arguments{
\item{ct_df_1}{a data frame with Ct data, 1 row per well.

Must have columns Ct, Dilution. 

Assumes data are only for 1 probe, i.e. all values in ct_df_1 are fit with the same slope.}

\item{formula}{formula to use for log-log regression fit. 

Default value assumes multiple biological replicates Ct ~ log2(Dilution) + BioRep.
If only a single Biological Replicate, change to Ct ~ log2(Dilution).}
}
\value{
data frame with 1 single row, and columns: efficiency, efficiency.sd, r.squared
}
\description{
Note efficiency is given in ratio, not per cent; multiply by 100 for that.
}
\seealso{
est_efficiency
}
