% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_efficiency.R
\name{est_efficiency_bytargetid}
\alias{est_efficiency_bytargetid}
\title{Calibrate multiple probes by calculating detection efficiency and R squared}
\usage{
est_efficiency_bytargetid(
  cq_df,
  formula = Cq ~ log2(Dilution) + BioRep,
  usetypes = "+RT"
)
}
\arguments{
\item{cq_df}{a data frame with Cq (quantification cycle) data, 1 row per well

Must have columns Type, target_id, Cq, Dilution.
Only Type=="+RT" columns are used.}

\item{formula}{formula to use for log-log regression fit.

Default value assumes multiple biological replicates,
Cq ~ log2(Dilution) + BioRep.

If only a single Biological Replicate, change to Cq ~ log2(Dilution).
If multiple sample_ids, change to Cq ~ log2(Dilution) + sample_id.

See ?formula for background and help.}

\item{usetypes}{Type column values to use, default "+RT" for RT-qPCR.

By default, this includes only reverse-transcribed values in the efficiency
estimation, so excludes negative controls such as no-template and no-RT.

To skip this filtering step, set usetypes=NA.}
}
\value{
data frame with columns:
target_id, efficiency, efficiency.sd, r.squared.
}
\description{
See calibration vignette for example of usage.
}
\details{
Note efficiency is given in ratio, not per cent; multiply by 100 for that.
}
\seealso{
est_efficiency
}
