% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_clean.R
\docType{data}
\name{ex_clean}
\alias{ex_clean}
\title{Example small cleaned qPCR data set}
\format{
A tibble with 69 rows and 5 variables:
\describe{
\item{treatment}{Three fictional treatment conditions, where CTRL is the
negative control}
\item{bio_rep}{Number to denote which biological replicate the value
belongs to}
\item{tech_rep}{Numbered technical replicates within a biological
replicate}
\item{primer_pair}{Three fictional genes targeted for qPCR, where gene_hk
is the housekeeping gene}
\item{cq_values}{Number of PCR cycles need to cross a fluorescence
threshold}
}
}
\source{
Fictional data set created in excel
}
\usage{
ex_clean
}
\description{
A fictional data set of randomly generated Cq values. Any outliers have been
removed with \code{qpcr_clean}, using a threshold of 1.
}
\keyword{datasets}
