% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_qpcr}
\alias{fake_qpcr}
\title{Example small qPCR data set}
\format{
A tibble with 81 rows and 5 variables:
\describe{
\item{treatment}{Three fictional treatment conditions, where CTRL is the
negative control}
\item{bio_rep}{Number to denote which biological replicate the value
belongs to}
\item{tech_rep}{Numbered technical replicates within a biological
replicate}
\item{primer_pair}{Three fictional genes targeted for qPCR, where gene_hk
is the housekeeping gene}
\item{cq_values}{Number of PCR cycles need to cross a fluorescence
threshold}
}
}
\source{
Fictional data set created in excel
}
\usage{
fake_qpcr
}
\description{
A fictional data set of randomly generated Cq values. Contains all possible
outlier cases that qpcr_clean removes.
}
\keyword{datasets}
