% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_melt_curve_functions.R
\name{getdRdTall}
\alias{getdRdTall}
\title{Calculate dR/dT for melt curves of every well in a plate}
\usage{
getdRdTall(platemelt, method = "spline", ...)
}
\arguments{
\item{platemelt}{data frame describing melt curves, including variables
Well, Temperature, Fluor (fluorescence value).}

\item{method}{to use for smoothing:

  "spline" default, uses smoothing spline stats::smooth.spline.
  
  "diff" base::diff for lagged difference}

\item{...}{other arguments to pass to smoothing method.}
}
\value{
platemelt with additional column dRdT.
}
\description{
Calculate dR/dT for melt curves of every well in a plate
}
\seealso{
Other melt_curve_functions: 
\code{\link{getdRdT}()}
}
\concept{melt_curve_functions}
