% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{log_plot_helpers}
\alias{log_plot_helpers}
\alias{scale_x_log2nice}
\alias{scale_x_log10nice}
\alias{scale_y_log2nice}
\alias{scale_y_log10nice}
\alias{scale_loglog10}
\title{Nice tick labels for logarithmic axes in ggplot2.}
\usage{
scale_x_log2nice(name = ggplot2::waiver(), omag = seq(-10, 10),
  scilabels = FALSE, ...)

scale_x_log10nice(name = ggplot2::waiver(), omag = seq(-10, 10),
  scilabels = FALSE, ...)

scale_y_log2nice(name = ggplot2::waiver(), omag = seq(-10, 10),
  scilabels = FALSE, ...)

scale_y_log10nice(name = ggplot2::waiver(), omag = seq(-10, 10),
  scilabels = FALSE, ...)

scale_loglog10(...)
}
\arguments{
\item{name}{axis name.}

\item{omag}{orders of magnitude or fold-changes for axis labels, 
usually an integer sequence.}

\item{scilabels}{display labels in scientific format, e.g. \eqn{10^2} vs 100.}

\item{...}{other arguments to continuous_scale().}
}
\value{
a ggproto object as output by continuous_scale().
}
\description{
These functions are wrappers for scale_x_log10, etc., that specify
axis tick labels as powers of 10 and powers of 2 with default
nice format or easy scientific notation.
}
\section{Functions}{
\itemize{
\item \code{scale_x_log2nice}: plot x axis on log2-scale with nice defaults

\item \code{scale_x_log10nice}: plot x axis on log10-scale with nice defaults

\item \code{scale_y_log2nice}: plot y axis on log2-scale with nice defaults

\item \code{scale_y_log10nice}: plot y axis on log10-scale with nice defaults

\item \code{scale_loglog10}: plot x AND y axes on log10-scale with nice defaults
}}

