% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{getNormCq}
\alias{getNormCq}
\alias{normalizeqPCR}
\alias{normaliseqPCR}
\title{Normalize cycle count (log2-fold) data within SampleID}
\usage{
getNormCq(
  cq_df,
  value = "Cq",
  normTargetIDs = "ALG9",
  probename = "TargetID",
  normby.function = median
)

normalizeqPCR(
  cq_df,
  value = "Cq",
  normTargetIDs = "ALG9",
  probename = "TargetID"
)

normaliseqPCR(
  cq_df,
  value = "Cq",
  normTargetIDs = "ALG9",
  probename = "TargetID"
)
}
\arguments{
\item{cq_df}{a data frame containing columns "SampleID", value (default Cq)
and probename (default TargetID). Crucially, SampleID should be the same
for different technical replicates measuring identical reactions in
different wells of the plate, but differ for different biological and
experimental replicates.}

\item{value}{the column name of the value that will be normalized}

\item{normTargetIDs}{names of PCR probes (or primer sets) to normalize by,
i.e. reference genes}

\item{probename}{the column name for probe sets}

\item{normby.function}{Function to use to calculate the value to
normalise by on log2/Cq scale.
Default value is median, alternatively could use mean.}
}
\value{
data frame like cq_df with three additional columns:

\tabular{ll}{
  norm.by       \tab the median value of the reference probes  \cr
  Value.norm    \tab the normalized value, \eqn{\Delta Cq} \cr
  Value.normexp \tab the normalized ratio, \eqn{2^(-\Delta Cq)}
  }
}
\description{
Normalize cycle count (log2-fold) data within SampleID
}
\section{Functions}{
\itemize{
\item \code{getNormCq}: get the median value of a set of normalization
(reference) probes, for each sample.

\item \code{normaliseqPCR}: Normalise cycle count (log2-fold) data within
SampleID. Synonym for normalizeqPCR.
}}

