% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_dcq.R
\name{qpcr_dcq}
\alias{qpcr_dcq}
\title{Calculate dCq values}
\usage{
qpcr_dcq(.data, cq, primer_pair, housekeeping, ...)
}
\arguments{
\item{...}{}
}
\value{
A tibble
}
\description{
This function subtracts the housekeeping gene Cq values from the experimental genes Cq values.
}
