% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_summary.R
\name{qpcr_summary}
\alias{qpcr_summary}
\title{Calculate mean and sd}
\usage{
qpcr_summary(data, to_summarise, ...)
}
\arguments{
\item{data}{The dataset you want to summarize, a tibble.}

\item{to_summarise}{The column you want to apply the function to, for example
the fold change values.}

\item{...}{The variables you want to group your data by. Mean and SD will be
calculated for every unique combination of these variables.}
}
\value{
A tibble
}
\description{
This function calculates the mean and standard deviation.
}
