% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr_data.R
\name{read_lc1536_1colour_cq}
\alias{read_lc1536_1colour_cq}
\title{Reads quantification cycle (cq) data in 1 colour from Roche Lightcycler 1536}
\usage{
read_lc1536_1colour_cq(
  filename,
  col_names = c("wellno", "well_lc1536", "include", "pipctrl", "target", "sample_info",
    "edited", "call", "cq", "slope", "epf"),
  col_types = "iclcccllnnn",
  cols_keep = c("well_lc1536", "cq"),
  skip = 4,
  ...
)
}
\arguments{
\item{filename}{file name}

\item{col_names}{names to give to columns}

\item{col_types}{data types of columns}

\item{cols_keep}{names of input columns to actually keep}

\item{skip}{number of lines to skip, defaults to 2}

\item{...}{other arguments to pass to read_tsv, if needed}
}
\value{
tibble containing cq data, with default column names:

  well_lc1536: the well of the plate in lc1536 format, Aa01, Aa02, ..., Hd48,
  as output make_row_names_echo1536()

  well: the well of the plate in standard/echo format A1, A2, ..., AF48, 

  cq: quantification cycle, numberic value
}
\description{
This is the data from "export in text format" from the analysis tab in the
Lightcycler software. That software calls cq, "Cp".
}
\details{
This function is a wrapper around `readr::read_tsv`, also selects only some
columns to keep, and returns the well names in multiple formats. 
Multiple formats are needed because the row names for lc1536 are 
Aa, Ab, Ac, Ac, Ba, ..., Hd, as output by `make_row_names_lc1536()`.
But the row names for Echo and other more standard systems are 
A, B, ..., Z, AA, AB, ..., AF, as output by `make_row_names_echo1536()`

This function outputs both, to be more flexible.
}
\seealso{
read_lightcycler_1colour_raw
}
