% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr_data.R
\name{read_lightcycler_1colour_cq}
\alias{read_lightcycler_1colour_cq}
\title{Reads quantification cycle (cq) data in 1 colour from Roche Lightcyclers}
\usage{
read_lightcycler_1colour_cq(
  filename,
  skip = 2,
  col_names = c("include", "color", "well", "sample_info", "cq", "concentration",
    "standard", "status"),
  col_types = "liccddil",
  ...
)
}
\arguments{
\item{filename}{file name}

\item{skip}{number of lines to skip, defaults to 2}

\item{col_names}{names to give to columns}

\item{col_types}{data types of columns}

\item{...}{other arguments to pass to read_tsv, if needed}
}
\value{
tibble containing cq data
}
\description{
This is the data from "export in text format" from the analysis tab in the
Lightcycler software. That software calls cq, "Cp".
}
\details{
This function is a thin wrapper around readr::read_tsv.
}
\examples{
read_lightcycler_1colour_cq(system.file("extdata/Edward_qPCR_Nrd1_calibration_2019-02-02_Ct.txt.gz", 
                                                 package = "tidyqpcr"))

}
\seealso{
read_lightcycler_1colour_raw
}
