% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr_data.R
\name{read_lightcycler_1colour_raw}
\alias{read_lightcycler_1colour_raw}
\title{Reads raw text-format fluorescence data in 1 colour from Roche Lightcyclers}
\usage{
read_lightcycler_1colour_raw(
  filename,
  skip = 2,
  col_names = c("well", "sample_info", "program_no", "segment_no", "cycle", "time",
    "temperature", "fluor_raw"),
  col_types = "ccffinnn",
  ...
)
}
\arguments{
\item{filename}{file name}

\item{skip}{number of lines to skip, defaults to 2}

\item{col_names}{names to give to columns}

\item{col_types}{data types of columns}

\item{...}{other arguments to pass to read_tsv, if needed}
}
\value{
tibble containing raw data, with default column names:

  well: the well of the plate, e.g. A1

  sample_info: this is the "Sample" field entered in lightcycler software,
  defaults to "Sample X"

  program_no: the number of the cycler program, for 2-step PCR defaults to 1
  = melt, 2 = amplify, 3 = melt.

  segment_no: the number of the segment of the cycler program, e.g.
  hold/raise/lower temperature

  cycle: the cycle number, for programs with repeated cycles (i.e.
  amplification)

  time: the time of fluorescence reading acquisition (in what units???)

  temperature: the temperature of the block at fluorescence acquisition

  fluor_raw: the raw fluorescence reading in "arbitrary units". For SYBR safe, this
  would be 483nm excitation, 533nm emission.
}
\description{
This is the data from "export in text format" from the Lightcycler software.
The other data format, .ixo, can be converted to .txt format by the
Lightcycler software.
}
\details{
This function is a thin wrapper around readr::read_tsv.
}
\seealso{
read_lightcycler_1colour_cq
}
