% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyqpcr-package.R
\docType{package}
\name{tidyqpcr-package}
\alias{tidyqpcr}
\alias{tidyqpcr-package}
\title{tidyqpcr: Quantitative PCR Analysis with the Tidyverse}
\description{
This package is for reproducible quantitative PCR (qPCR) analysis using packages from the tidyverse, notably dplyr and ggplot2. It normalizes (by ddCq), summarizes, and plots pre-calculated Cq data, and plots raw amplification and melt curves from Roche lightcycler machines. It does NOT (yet) calculate Cq data from amplification curves.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ewallace/tidyqpcr}
  \item \url{https://ewallace.github.io/tidyqpcr/}
  \item Report bugs at \url{https://github.com/ewallace/tidyqpcr/issues}
}

}
\author{
\strong{Maintainer}: Edward Wallace \email{Edward.Wallace@ed.ac.uk}

Other contributors:
\itemize{
  \item Sam Haynes \email{samuel.haynes10@gmail.com} [contributor]
}

}
\keyword{internal}
