% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deltacq.R
\name{calculate_deltacq_bysampleid}
\alias{calculate_deltacq_bysampleid}
\title{Calculate delta cq (\eqn{\Delta Cq}) to normalize quantification cycle
(log2-fold) data within sample_id.}
\usage{
calculate_deltacq_bysampleid(cq_df, ref_target_ids, norm_function = median)
}
\arguments{
\item{cq_df}{a data frame containing columns `sample_id`, value_name (default
`cq`) and tid_name (default `target_id`). Crucially, sample_id should be
the same for different technical replicates measuring identical reactions
in different wells of the plate, but differ for different biological and
experimental replicates. See tidyqpcr vignettes for examples.}

\item{ref_target_ids}{names of PCR probes (or primer sets) to normalize by,
i.e. reference genes}

\item{norm_function}{Function to use to calculate the value to normalize by
on given scale. Default is median, alternatively could use mean.}
}
\value{
data frame like cq_df with three additional columns:

  \tabular{ll}{ ref_cq    \tab summary (median/mean) cq value for reference
  target ids \cr delta_cq  \tab normalized value, \eqn{\Delta Cq} \cr
  rel_abund \tab normalized ratio, \eqn{2^(-\Delta Cq)} }
}
\description{
Calculate delta cq (\eqn{\Delta Cq}) to normalize quantification cycle
(log2-fold) data within sample_id.
}
