% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_efficiency.R
\name{calculate_efficiency_bytargetid}
\alias{calculate_efficiency_bytargetid}
\title{Calibrate multiple probes by calculating detection efficiency and R squared}
\usage{
calculate_efficiency_bytargetid(
  cq_df,
  formula = cq ~ log2(dilution) + BioRep,
  use_prep_types = "+RT"
)
}
\arguments{
\item{cq_df}{a data frame with cq (quantification cycle) data, 1 row per well

Must have columns prep_type, target_id, cq, dilution.
Only prep_type=="+RT" columns are used.}

\item{formula}{formula to use for log-log regression fit.

Default value assumes multiple biological replicates,
cq ~ log2(dilution) + BioRep.

If only a single Biological Replicate, change to cq ~ log2(dilution).
If multiple sample_ids, change to cq ~ log2(dilution) + sample_id.

See ?formula for background and help.}

\item{use_prep_types}{prep_type column values to use, default "+RT" for RT-qPCR.

By default, this includes only reverse-transcribed values in the efficiency
estimation, so excludes negative controls such as no-template and no-RT.

To skip this filtering step, set use_prep_types=NA.}
}
\value{
data frame with columns:
target_id, efficiency, efficiency.sd, r.squared.
}
\description{
See calibration vignette for example of usage.
}
\details{
Note efficiency is given in ratio, not per cent; multiply by 100 for that.
}
\seealso{
calculate_efficiency
}
