% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_blank_plate}
\alias{create_blank_plate}
\alias{create_blank_plate_96well}
\alias{create_blank_plate_1536well}
\title{Create a blank plate template as a tibble (with helper functions for common plate sizes)}
\usage{
create_blank_plate(well_row = LETTERS[1:16], well_col = 1:24)

create_blank_plate_96well()

create_blank_plate_1536well(
  well_row = make_row_names_lc1536(),
  well_col = 1:48
)
}
\arguments{
\item{well_row}{Vector of Row labels, usually LETTERS}

\item{well_col}{Vector of Column labels, usually numbers}
}
\value{
tibble (data frame) with columns well_row, well_col, well. This
  contains all pairwise combinations of well_row and well_col, as well as
  individual well names. Both well_row and well_col are coerced to factors
  (even if well_col is supplied as numbers), to ensure order is consistent.

  However, well is a character vector as that is the default behaviour of
  "unite", and display order doesn't matter.

  Default value describes a full 384-well plate.
}
\description{
For more help, examples and explanations, see the plate setup vignette:
\code{vignette("platesetup_vignette", package = "tidyqpcr")}
}
\section{Functions}{
\itemize{
\item \code{create_blank_plate_96well}: create blank 96-well plate

\item \code{create_blank_plate_1536well}: create blank 1536-well plate
}}

\examples{
create_blank_plate(well_row=LETTERS[1:2],well_col=1:3)

create_blank_plate_96well()

create_blank_plate_1536well()

# create blank 96-well plate with empty edge wells

create_blank_plate(well_row=LETTERS[2:7], well_col=2:11)

# create blank 1536-well plate with empty edge wells

full_plate_row_names <- make_row_names_lc1536()

create_blank_plate(well_row=full_plate_row_names[2:31], well_col=2:47)

}
\seealso{
Other plate creation functions: 
\code{\link{create_colkey_4diln_2ctrl_in_24}()},
\code{\link{create_colkey_6_in_24}()},
\code{\link{create_colkey_6diln_2ctrl_in_24}()},
\code{\link{create_rowkey_4_in_16}()},
\code{\link{create_rowkey_8_in_16_plain}()},
\code{\link{display_plate_qpcr}()},
\code{\link{display_plate_value}()},
\code{\link{display_plate}()},
\code{\link{label_plate_rowcol}()},
\code{\link{make_row_names_echo1536}()},
\code{\link{make_row_names_lc1536}()}
}
\concept{plate creation functions}
