% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{create_blank_plate}
\alias{create_blank_plate}
\alias{create_blank_plate_96well}
\alias{make_row_names_LC_1536}
\alias{make_row_names_Echo_1536}
\alias{create_blank_plate_1536well}
\title{Create a blank plate template as a tibble}
\usage{
create_blank_plate(WellR = LETTERS[1:16], WellC = 1:24)

create_blank_plate_96well()

make_row_names_LC_1536()

make_row_names_Echo_1536()

create_blank_plate_1536well(WellR = make_row_names_LC_1536(),
  WellC = 1:48)
}
\arguments{
\item{WellR}{Vector of Row labels, usually LETTERS}

\item{WellC}{Vector of Column labels, usually numbers}
}
\value{
tibble (data frame) with columns WellR, WellC, Well. This contains
  all pairwise combinations of WellR and WellC, as well as individual Well
  names. Both WellR and WellC are coerced to factors (even if WellC
  is supplied as numbers), to ensure order is consistent.
  
  However, Well is a character vector as that is the default behaviour of 
  "unite", and display order doesn't matter.
  
  Default value describes a full 384-well plate.
}
\description{
Create a blank plate template as a tibble
}
\section{Functions}{
\itemize{
\item \code{create_blank_plate_96well}: create blank 96-well plate

\item \code{make_row_names_LC_1536}: Row names for 1536-well plates on Lightcycler 1536 Aa,Ab,Ac,Ad,Ba,...,Hd.

\item \code{make_row_names_Echo_1536}: Row names for 1536-well plates on Labcyte Echo A,B,...,Z,AA,AB,...,AF.

\item \code{create_blank_plate_1536well}: create blank 1536-well plate
}}

\examples{
create_blank_plate(WellR=LETTERS[1:2],WellC=1:3)
create_blank_plate_96well()
}
\seealso{
Other plate creation functions: \code{\link{create_colkey_4dilutions_mRTNT_in24}},
  \code{\link{create_colkey_6dilutions_mRTNT_in24}},
  \code{\link{create_colkey_6in24}},
  \code{\link{create_rowkey_4in16}},
  \code{\link{create_rowkey_8in16_plain}},
  \code{\link{display_plate}},
  \code{\link{label_plate_rowcol}}
}
\concept{plate creation functions}
