% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_melt_curve_functions.R
\name{getdRdT}
\alias{getdRdT}
\title{Calculate dR/dT for a melt curve of signal R vs temperature T}
\usage{
getdRdT(TT, RR, method = c("spline", "diff"), ...)
}
\arguments{
\item{TT}{Temperature.}

\item{RR}{Signal, assumed fluorescence signal matched to TT.}

\item{method}{to use, defaulting to a smoothing spline.}

\item{...}{other arguments to pass to smoothing method.}
}
\value{
estimated first derivative of RR with respect to TT, 
numeric vector of same length as RR.
}
\description{
Calculate dR/dT for a melt curve of signal R vs temperature T
}
\seealso{
Other melt_curve_functions: 
\code{\link{getdRdTall}()}
}
\concept{melt_curve_functions}
