% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_functions.R
\name{getNormCt}
\alias{getNormCt}
\alias{normalizeqPCR}
\alias{normaliseqPCR}
\title{Normalize cycle count (log2-fold) data within Sample}
\usage{
getNormCt(ct_df, value = "Ct", normProbes = "ALG9", probename = "Probe")

normalizeqPCR(ct_df, value = "Ct", normProbes = "ALG9", probename = "Probe")

normaliseqPCR(ct_df, value = "Ct", normProbes = "ALG9", probename = "Probe")
}
\arguments{
\item{ct_df}{a data frame containing columns "Sample", value (default Ct) and
probe (default Probe). Crucially, Sample name should be the same for
different technical replicates measuring identical reactions in different
wells of the plate, but differ for different biological and experimental 
replicates.}

\item{value}{the column name of the value that will be normalized}

\item{normProbes}{names of PCR probes (or primer sets) to normalize by, i.e.
reference genes}

\item{probename}{the column name for probe sets}
}
\value{
data frame like ct_df with three additional columns:

\tabular{ll}{
  norm.by       \tab the median value of the reference probes  \cr
  Value.norm    \tab the normalized value, \eqn{\Delta Ct} \cr
  Value.normexp \tab the normalized ratio, \eqn{2^(-\Delta Ct)}
  }
}
\description{
Normalize cycle count (log2-fold) data within Sample
}
\section{Functions}{
\itemize{
\item \code{getNormCt}: get the median value of a set of normalization
(reference) probes, for each sample.

\item \code{normaliseqPCR}: Normalise cycle count (log2-fold) data within Sample.
Synonym for normalizeqPCR.
}}

