% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_avg_techrep.R
\name{qpcr_avg_techrep}
\alias{qpcr_avg_techrep}
\title{Average technical replicates}
\usage{
qpcr_avg_techrep(.data, cq, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
This function calculates the average of every technical replicate.
na.rm is set at TRUE by default.
}
\details{
To do: control na.rm from funcion.
}
