% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_avg_techrep.R
\name{qpcr_avg_techrep}
\alias{qpcr_avg_techrep}
\title{Average technical replicates}
\usage{
qpcr_avg_techrep(.data, cq, ...)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{cq}{Unquoted expression. The name of the column containing the Cq
values.}

\item{...}{Tidy-select. One or more unquoted expressions separated by commas.
All the names of columns that denote your experimental variables (and
biological replicates). The average will be taken for every unique
combination of values in the supplied columns. Be sure to exclude the
columns containing the Cq values or technical replicates.}
}
\value{
Returns the same type as the input (e.g. a data frame or tibble). All
columns supplied in ... will be returned. Creates a new column for the
averaged Cq values with the name "cq".
}
\description{
This function calculates the average of every technical replicate.
\code{mean} function argument na.rm is set to TRUE by default.
}
\examples{
avg_data <- qpcr_avg_techrep(ex_clean,
                             cq = cq_values,
                             treatment, primer_pair, bio_rep)
}
