% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_clean.R
\name{qpcr_clean}
\alias{qpcr_clean}
\title{Remove outliers}
\usage{
qpcr_clean(.data, cq, threshold, ...)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{cq}{Unquoted expression. The name of the column containing the Cq
values.}

\item{threshold}{A numeric. The maximum allowed deviation from the median.}

\item{...}{Tidy-select. One or more unquoted expressions separated by commas.
These are the names of all other columns that are not the Cq values or
denote technical replicates. They will be used to make groups, so
calculations will be made for each unique combination of variables.}
}
\value{
Returns the same type as the input (e.g. a data frame or tibble).
Rows that are deemed outliers are removed from the output. If you want to
inspect these rows see \code{\link{qpcr_outlier_context}}.
}
\description{
\code{qpcr_clean} removes outliers from technical qPCR replicates. These
outliers can occur because of inaccurate pipetting, pipetting in the wrong
well, seal detachment, etc. The function removes them based on the deviation
from the median Cq value. See details for exact rules of outlier removal.
}
\details{
The rules for what is an outlier and what is not are as follows:
\enumerate{
\item If only one Cq value is present (i.e. the other replicates failed to
produce a Cq value), it will be removed.
\item If only two Cq values are present, they need to be less than the
threshold apart.
\item For three or more technical replicates:
\itemize{
\item If the absolute distance between a Cq value and the median Cq
is greater than the set threshold, than this value will be
removed.
\item If all Cq values within a technical replicate are more than
the threshold apart, they will all be removed. } }
}
