% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_dcq.R
\name{qpcr_dcq}
\alias{qpcr_dcq}
\alias{qpcr_dct}
\alias{qpcr_dcp}
\title{Calculate dCq values}
\usage{
qpcr_dcq(.data, cq, pcr_target, housekeeping)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{cq}{Unquoted expression. Give the name of the column containing the
(averaged) Cq values. If you used the qpcr_avg_techrep function this column
is called 'cq'.}

\item{pcr_target}{Unquoted expression. Supply the name of the column that
denotes which gene was the target of the qPCR.}

\item{housekeeping}{Quoted expression. Give the value of your housekeeping
gene as it occurs in your pcr_target column.}
}
\value{
Returns the same type as the input (e.g. a data frame or tibble).
Creates 2 new columns. "cq_hk" moves the housekeeping Cq values into its
own column. "dcq" contains the calculated dCq values.
}
\description{
\code{qpcr_dcq} subtracts the housekeeping gene Cq values from the
experimental genes Cq values.
}
\examples{
dcq_values <- qpcr_dcq(ex_avg,
                       cq = cq,
                       pcr_target = primer_pair,
                       housekeeping = "gene_hk")
}
