% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_dcq.R
\name{qpcr_dcq}
\alias{qpcr_dcq}
\title{Calculate dCq values}
\usage{
qpcr_dcq(.data, cq, primer_pair, housekeeping, ...)
}
\arguments{
\item{.data}{The name of the dataset containing the average Cq values.}

\item{cq}{Give the name of the column containing the (averaged) Cq values. If
you used the qpcr_avg_techrep function this column is called 'cq'.}

\item{primer_pair}{Supplies the name of the column that denotes which gene
was the target of the PCR.}

\item{housekeeping}{Give the value of your housekeeping gene, in quotes, as
it occurs in your primer_pair column.}

\item{...}{All other column names excluding columns for: cq values, technical
replicates and primers used.}
}
\value{
A tibble
}
\description{
This function subtracts the housekeeping gene Cq values from the experimental
genes Cq values.
}
