% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqcr_ddcq.R
\name{qpcr_ddcq}
\alias{qpcr_ddcq}
\title{Calculate ddCq values}
\usage{
qpcr_ddcq(.data, dcq = dcq, treatment, untreated, primer_pair)
}
\arguments{
\item{primer_pair}{}
}
\value{
A tibble
}
\description{
This function subtracts the dCq control from the dCq treatments.
It also calculates the fold change.
}
