% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_outlier_context.R
\name{qpcr_outlier_context}
\alias{qpcr_outlier_context}
\title{Show outlier context}
\usage{
qpcr_outlier_context(raw_data, clean_data, cq_values, tech_rep, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Shows all the technical replicates that contain at least one outlier.
This function takes the raw data and the output from qpcr_clean() as input.
}
