% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_outlier_context.R
\name{qpcr_outlier_context}
\alias{qpcr_outlier_context}
\title{Show outlier context}
\usage{
qpcr_outlier_context(raw_data, clean_data, cq_values, tech_rep, ...)
}
\arguments{
\item{raw_data}{Here you supply the raw unfiltered data.}

\item{clean_data}{Here you supply the cleaned data from the qpcr_clean
function.}

\item{cq_values}{Give name of the column containing the cq values.}

\item{tech_rep}{Give the name of the column containing the technical
replicate information.}

\item{...}{All other column names, excluding the columns containing the Cq
values or technical replicates. Give them unquoted and separated by a
comma.}
}
\value{
A tibble
}
\description{
Shows all the technical replicates that contain at least one outlier.
This function takes the raw data and the output from qpcr_clean() as input.
}
