% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_summary.R
\name{qpcr_summary}
\alias{qpcr_summary}
\title{Calculate mean and sd}
\usage{
qpcr_summary(data, to_summarise, ...)
}
\arguments{
\item{...}{}
}
\value{
A tibble
}
\description{
This function calculates the mean and standard deviation.
}
